установила вебпак, разбила код на компоненты
