+++ public/index.html
@@ -7,7 +7,7 @@
     <title>Big Trip</title>
 
     <link href="https://fonts.googleapis.com/css?family=Montserrat:400,500,600,700,800&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="css/style.css">
+    <link rel="stylesheet" href="./css/style.css">
   </head>
 
   <body class="page-body">
@@ -41,6 +41,6 @@
           </section>
       </div>
     </main>
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
