+++ src/components/available-offer.js
@@ -1,0 +1,14 @@
+const createAvailableOfferTemplate = () => {
+  return (
+    `<div class="event__offer-selector">
+      <input class="event__offer-checkbox  visually-hidden" id="event-offer-luggage-1" type="checkbox" name="event-offer-luggage" checked>
+      <label class="event__offer-label" for="event-offer-luggage-1">
+        <span class="event__offer-title">Add luggage</span>
+        &plus;&euro;&nbsp;
+        <span class="event__offer-price">30</span>
+      </label>
+    </div>`
+  );
+};
+
+export {createAvailableOfferTemplate};
