+++ src/components/event-type.js
@@ -1,0 +1,10 @@
+const createEventTypeTemplate = () => {
+  return (
+    `<div class="event__type-item">
+      <input id="event-type-taxi-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="taxi">
+      <label class="event__type-label  event__type-label--taxi" for="event-type-taxi-1">Taxi</label>
+    </div>`
+  );
+};
+
+export {createEventTypeTemplate};
