+++ src/components/trip-days.js
@@ -1,0 +1,8 @@
+const createDayListTemplate = () => {
+  return (
+    `<ul class="trip-days">
+    </ul>`
+  );
+};
+
+export {createDayListTemplate};
